refactor : ShopAddress 경로 이동
