+++ mealkitary-application/src/test/kotlin/com/mealkitary/shop/application/service/GetShopServiceTest.kt
@@ -5,10 +5,10 @@ import com.mealkitary.common.model.Coordinates
 import com.mealkitary.shop.application.port.input.ShopResponse
 import com.mealkitary.shop.application.port.output.LoadShopPort
 import com.mealkitary.shop.domain.shop.Shop
+import com.mealkitary.shop.domain.shop.ShopAddress
 import com.mealkitary.shop.domain.shop.ShopBusinessNumber
 import com.mealkitary.shop.domain.shop.ShopStatus
 import com.mealkitary.shop.domain.shop.ShopTitle
-import com.mealkitary.shop.domain.shop.address.ShopAddress
 import io.kotest.core.spec.style.AnnotationSpec
 import io.kotest.matchers.shouldBe
 import io.mockk.every
