+++ mealkitary-application/src/test/kotlin/com/mealkitary/shop/application/service/RegisterShopServiceTest.kt
@@ -6,10 +6,10 @@ import com.mealkitary.shop.application.port.input.RegisterShopRequest
 import com.mealkitary.shop.application.port.output.SaveShopPort
 import com.mealkitary.shop.domain.product.Product
 import com.mealkitary.shop.domain.shop.Shop
+import com.mealkitary.shop.domain.shop.ShopAddress
 import com.mealkitary.shop.domain.shop.ShopBusinessNumber
 import com.mealkitary.shop.domain.shop.ShopStatus
 import com.mealkitary.shop.domain.shop.ShopTitle
-import com.mealkitary.shop.domain.shop.address.ShopAddress
 import com.mealkitary.shop.domain.shop.factory.AddressResolver
 import com.mealkitary.shop.domain.shop.factory.ShopBusinessNumberValidator
 import com.mealkitary.shop.domain.shop.factory.ShopFactory
