+++ mealkitary-domain/src/main/kotlin/com/mealkitary/shop/domain/shop/ShopAddress.kt
@@ -1,4 +1,4 @@
-package com.mealkitary.shop.domain.shop.address
+package com.mealkitary.shop.domain.shop
 
 import com.mealkitary.common.model.Address
 import com.mealkitary.common.model.Coordinates
