+++ mealkitary-domain/src/main/kotlin/com/mealkitary/shop/domain/shop/factory/AddressResolver.kt
@@ -1,6 +1,6 @@
 package com.mealkitary.shop.domain.shop.factory
 
-import com.mealkitary.shop.domain.shop.address.ShopAddress
+import com.mealkitary.shop.domain.shop.ShopAddress
 
 interface AddressResolver {
