+++ mealkitary-domain/src/test/kotlin/com/mealkitary/shop/domain/shop/factory/ShopFactoryTest.kt
@@ -2,7 +2,7 @@ package com.mealkitary.shop.domain.shop.factory
 
 import com.mealkitary.common.model.Address
 import com.mealkitary.common.model.Coordinates
-import com.mealkitary.shop.domain.shop.address.ShopAddress
+import com.mealkitary.shop.domain.shop.ShopAddress
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.AnnotationSpec
 import io.kotest.matchers.shouldBe
