+++ mealkitary-domain/src/testFixtures/kotlin/data/ShopTestData.kt
@@ -4,10 +4,10 @@ import com.mealkitary.common.model.Address
 import com.mealkitary.common.model.Coordinates
 import com.mealkitary.shop.domain.product.Product
 import com.mealkitary.shop.domain.shop.Shop
+import com.mealkitary.shop.domain.shop.ShopAddress
 import com.mealkitary.shop.domain.shop.ShopBusinessNumber
 import com.mealkitary.shop.domain.shop.ShopStatus
 import com.mealkitary.shop.domain.shop.ShopTitle
-import com.mealkitary.shop.domain.shop.address.ShopAddress
 import data.ProductTestData.Companion.defaultProduct
 import java.time.LocalTime
